+++ client-course-schedulizer/src/components/Tabs/tabComponents/CSVActions/CSVActions.tsx
@@ -3,13 +3,12 @@ import { Menu as MenuIcon } from "@material-ui/icons";
 import { ImportInputWrapper } from "components";
 import { bindMenu, bindTrigger, usePopupState } from "material-ui-popup-state/hooks";
 import React from "react";
-import { useExportCSV, useExportFullCSV } from "utilities";
+import { useExportFullCSV } from "utilities";
 import "./CSVActions.scss";
 
 /* Hamburger with options for the CSV */
 export const CSVActions = () => {
   const popupState = usePopupState({ popupId: "menu", variant: "popover" });
-  const onExportClick = useExportCSV();
   const onFullExportClick = useExportFullCSV();
 
   return (
@@ -35,11 +34,8 @@ export const CSVActions = () => {
         <ImportInputWrapper isAdditiveImport>
           <MenuItem button>ADD SCHEDULE</MenuItem>
         </ImportInputWrapper>
-        <MenuItem button onClick={onExportClick}>
-          EXPORT FINAL CSV
-        </MenuItem>
         <MenuItem button onClick={onFullExportClick}>
-          EXPORT DRAFT CSV
+          EXPORT CSV
         </MenuItem>
       </Menu>
     </>
