+++ client-course-schedulizer/src/components/Tabs/tabComponents/CSVActions/ExportButton/ExportButton.tsx
@@ -2,9 +2,7 @@ import { Button, ButtonProps } from "@material-ui/core";
 import React from "react";
 import { useExportCSV } from "utilities";
 
-/* A button that will export the current schedule as a CSV with the format specified by the registrar.
-  Can style with Mat UI props.
-*/
+// Deprecated, TODO: Remove code?
 export const ExportButton = (btnProps: ButtonProps) => {
   const onExportClick = useExportCSV();
